azure_function/.env/DiscountMovies/requirements.txt
azure-functions==1.0.0a5
azure-functions-worker==1.0.0a6
grpcio==1.14.2
grpcio-tools==1.14.2
protobuf==3.6.1
beautifulsoup4==4.6.3
nose==1.3.7
python-dateutil==2.7.5
six==1.11.0
pymongo==3.7.2